dashboard/backend/requirements.txt
sqlalchemy
fastapi
uvicorn
asyncpg
pydantic
python-dotenv
alembic
psycopg2-binary
fastapi-cache2
jinja2
pandas